fix tidy

+++ src/Compression/CompressionCodecGCD.cpp
@@ -211,7 +211,7 @@ void CompressionCodecGCD::doDecompressData(const char * source, UInt32 source_si
         throw Exception(ErrorCodes::CANNOT_DECOMPRESS, "Cannot decompress GCD-encoded data. File has wrong header");
 
     UInt8 bytes_to_skip = uncompressed_size % bytes_size;
-    chassert(bytes_to_skip == source[1]);
+    chassert(bytes_to_skip == static_cast<UInt8>(source[1]));
 
     UInt32 output_size = uncompressed_size - bytes_to_skip;
